+++ inst/examples/ex-describe.R
@@ -1,0 +1,12 @@
+## Create a data matrix
+X <- matrix(sample(0:9, 15, TRUE), nrow = 3, ncol = 5)
+
+## Add NA
+k <- sample(1:15, 3, FALSE)
+X[k] <- NA
+
+## Sparsity
+sparsity(X)
+
+## Quick description
+describe(X)